+++ CHANGELOG.md
@@ -12,10 +12,13 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 ### Added
 - #421 Added contract agreement view
 - #515 Service Unavailable Response on Notification failure
-- #536 Added import state PUBLISHED_TO_CX in frontend
+- #536 Added import state PUBLISHED_TO_CORE_SERVICES in frontend
 - #420 add /contracts api to fetch contract agreement information from EDC for assets
 - Added a step to the pull-request-backend.yml which checks if the pom.xml(root) properties have some versions ending with -SNAPSHOT
 - Added a PostConstruct method in PolicyStartUpConfig to allow Integration tests to run without errors in stack traces.
+- #536 added new ImportState to asset PUBLISHED_TO_CORE_SERVICES indicating edc assets and dtr shells were created for given asset
+- #536 added cron job responsible to publish assets in PUBLISHED_TO_CORE_SERVICES import state to edc and dtr
+
 ### Changed
 - Updated RELEASE.md to the latest release guide (added more steps)
 - #515 Fixed notification toast click area
@@ -23,6 +26,8 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #423 Moved errorMessages from investigation/alert to notification list
 - Updated COMPATIBILITY.md matrix adding release C-X 24.3 and 23.12
 - #536 rework /policies to respond with policies from the IRS policy store
+- #536 sync assets logic was adjusted to create IRS jobs only for assets that are not in TRANSIENT or IN_SYNC states
+- #536 Updated Arc42 documentation and user-manual with publish assets informations
 
 ### Removed
 - #625 Removed the header and breadcrumbs section from app layout